+++ pylatex/__init__.py
@@ -13,7 +13,7 @@ from .document import Document  # noqa
 from .math import Math  # noqa
 from .package import Package  # noqa
 from .section import Section, Subsection, Subsubsection  # noqa
-from .table import Table, Tabular  # noqa
+from .table import Table, MultiColumn, MultiRow, Tabular  # noqa
 from .pgfplots import TikZ, Axis, Plot  # noqa
 from .graphics import Figure, SubFigure, Plt  # noqa
 from .lists import Enumerate, Itemize, Description  # noqa